refactor: streamline user action opt-in handling and introduce new types for better structure
